+++ inject_runtime_pyproject.py
@@ -11,6 +11,18 @@ from hatchling.metadata.plugin.interface import MetadataHookInterface
 RUNTIME_PYPROJECT_PATH = "tesseract_core/runtime/meta/pyproject.toml"
 
 BASE_OPTIONAL_DEPS = {
+    "docs": [
+        "sphinx",
+        "sphinx_autodoc_typehints",
+        "furo",
+        "myst-parser",
+        "sphinx_click",
+        "autodoc_pydantic",
+        "sphinx_design",
+        "sphinx_copybutton",
+        "sphinxext_opengraph",
+        "tesseract-core[runtime]",
+    ],
     "dev": [
         "fastapi",
         "httpx",  # required by fastapi test client
@@ -24,9 +36,9 @@ BASE_OPTIONAL_DEPS = {
         "moto[server]",
         "aiobotocore>=2.19.0",  # without this pip dependency resolution fails
         "typeguard",
-        "detect-secrets[gibberish]",
         # also add all other extras here
         "tesseract-core[runtime]",
+        "tesseract-core[docs]",
     ],
     "runtime": [],
 }